fixed branching kernels test

+++ tests/branching_kernels/if_else_0_labels/if_else_0_labels_dcmpl-gfx1010.cl
@@ -1,7 +1,7 @@
 __kernel __attribute__((reqd_work_group_size(8, 8, 1)))
 void if_else_0_labels(int x, __global int *data, int y)
 {
-    int var0;
+    uint var0;
     uint var2;
     uint var3;
     var2 = get_global_id(0);
@@ -12,7 +12,7 @@ void if_else_0_labels(int x, __global int *data, int y)
     else {
         var3 = get_global_id(1);
         var2 = 1;
-        var0 = (get_global_id(1) * x) - (uint)y;
+        var0 = (get_global_id(1) * x) - y;
     }
     data[var2] = var0;
     data[var3] = x;
