+++ tests/branching_kernels/if_else_2_labels/if_else_2_labels_dcmpl-gfx1010.cl
@@ -1,7 +1,7 @@
 __kernel __attribute__((reqd_work_group_size(8, 8, 1)))
 void if_else_2_labels(int x, __global int *data, int y)
 {
-    int var0;
+    uint var0;
     uint var3;
     uint var4;
     var3 = get_global_id(0);
@@ -12,7 +12,7 @@ void if_else_2_labels(int x, __global int *data, int y)
     else {
         var4 = get_global_id(1);
         var3 = 1;
-        var0 = (get_global_id(1) * x) - (uint)y;
+        var0 = (get_global_id(1) * x) - y;
     }
     data[var3] = var0;
     data[var4] = x;
