+++ tests/branching_kernels/if_else_and_if_else_0_labels/if_else_and_if_else_0_labels_dcmpl-gfx1010.cl
@@ -3,7 +3,7 @@ void if_else_and_if_else_0_labels(int x, __global int *data, int y)
 {
     uint var1;
     uint var3;
-    int var5;
+    uint var5;
     uint var6;
     var3 = get_global_id(0);
     if (1 != var3) {
@@ -13,9 +13,9 @@ void if_else_and_if_else_0_labels(int x, __global int *data, int y)
     else {
         var1 = get_global_id(1);
         var3 = 1;
-        var5 = (get_global_id(1) * x) - (uint)y;
+        var5 = (get_global_id(1) * x) - y;
     }
-    var6 = x < y ? (ulong)x + get_global_id(2) : (ulong)y + var1;
+    var6 = x < y ? x + get_global_id(2) : y + var1;
     data[var3] = var5;
     data[get_global_id(2)] = var6;
     data[var1] = x;
