+++ tests/branching_kernels/if_else_in_if_of_if_else_gfx/if_else_in_if_of_if_else_gfx_dcmpl-gfx1030.cl
@@ -3,7 +3,7 @@ void if_else_in_if_of_if_else_1(int x, __global int *data, int y)
 {
     uint var4;
     uint var6;
-    int var8;
+    uint var8;
     var6 = get_global_id(0);
     if (1 != var6) {
         var8 = get_global_id(0) * y;
@@ -11,14 +11,14 @@ void if_else_in_if_of_if_else_1(int x, __global int *data, int y)
     }
     else {
         var4 = get_global_id(1);
-        data[get_global_id(0)] = (get_global_id(1) * x) - (uint)y;
+        data[get_global_id(0)] = (get_global_id(1) * x) - y;
         if (x >= y) {
-            var8 = (ulong)y + get_global_id(1);
+            var8 = y + get_global_id(1);
             var6 = get_global_id(2);
         }
         else {
             var6 = get_global_id(2);
-            var8 = (ulong)x + get_global_id(2);
+            var8 = x + get_global_id(2);
         }
     }
     data[var6] = var8;
@@ -28,7 +28,7 @@ void if_else_in_if_of_if_else_1(int x, __global int *data, int y)
 __kernel __attribute__((reqd_work_group_size(4, 4, 4)))
 void if_else_in_if_of_if_else_2(int x, __global int *data, int y)
 {
-    int var12;
+    uint var12;
     uint var5;
     var5 = get_global_id(0);
     if (1 != var5) {
@@ -36,12 +36,12 @@ void if_else_in_if_of_if_else_2(int x, __global int *data, int y)
     }
     else {
         var5 = get_global_id(1);
-        data[get_global_id(0)] = (get_global_id(1) * x) - (uint)y;
+        data[get_global_id(0)] = (get_global_id(1) * x) - y;
         if (var5 >= get_global_id(2)) {
-            var12 = (ulong)y + get_global_id(1);
+            var12 = y + get_global_id(1);
         }
         else {
-            var12 = (ulong)x + get_global_id(2);
+            var12 = x + get_global_id(2);
             var5 = get_global_id(2);
         }
     }
